+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 [<img height='36' style='border:0px;height:36px;' src='https://az743702.vo.msecnd.net/cdn/kofi2.png?v=2' border='0' alt='Buy Me a Coffee at ko-fi.com'/>](https://ko-fi.com/X8X41IE4T)
 
 ## 2.0.3
-
+* LuaLS library package:
+  * display `@deprecated` message on more removed/moved apis
 
 ## 2.0.2
 * LuaLS library package: