+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 [<img height='36' style='border:0px;height:36px;' src='https://az743702.vo.msecnd.net/cdn/kofi2.png?v=2' border='0' alt='Buy Me a Coffee at ko-fi.com'/>](https://ko-fi.com/X8X41IE4T)
 
 ## 2.0.6
+* VSCode:
+  * Migrate extension to ESM. Requires VSCode >=1.100
 * LuaLS library package:
   * remove `(exact)` from generated class `event_handler.events`
   * set `"Lua.language.fixIndent": false` by default due to bugs in LuaLS causing incorrect code deletions