+++ CHANGELOG.md
@@ -15,6 +15,8 @@
   * Omit `GuiStyle` from `AnyPrototype`, to reduce incorrect type deduction on data:extend's argument
   * `LuaObject`s now narrow on `object_name` without plugin assistance
   * include string builtin event name overloads for `on_event`, and fields in `event_handler.events`
+  * add deprecation tags on defines.inventory
+  * fix incorrectly generated type on defines.prototypes
 * Debugger:
   * fix incorrect error when remote.call is called with invalid arguments
   * fix incorrect nesting of values in Output events